+++ src/sync/helpers.js
@@ -1,0 +1,19 @@
+// @flow
+
+import type { ColumnName } from '..'
+import type { RawRecord } from '../RawRecord'
+
+export function addToRawSet(rawSet: ?string, value: string): string {
+  const array = rawSet ? rawSet.split(',') : []
+  const set = new Set(array)
+  set.add(value)
+  return Array.from(set).join(',')
+}
+
+// Mutates `rawRecord` to mark `columName` as modified for sync purposes
+export function setRawColumnChange(rawRecord: RawRecord, columnName: ColumnName): void {
+  if (rawRecord._status !== 'created') {
+    rawRecord._changed = addToRawSet(rawRecord._changed, columnName)
+    rawRecord._status = 'updated'
+  }
+}
